test: Refactor getTypicalAddressBook
