+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -1,8 +1,11 @@
 package seedu.address.testutil;
 
+import static seedu.address.testutil.TypicalPersons.getTypicalPersons;
+import static seedu.address.testutil.TypicalTasks.getTypicalTasks;
+
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Person;
-
+import seedu.address.model.task.Task;
 /**
  * A utility class to help with building Addressbook objects.
  * Example usage: <br>
@@ -30,5 +33,19 @@ public class AddressBookBuilder {
 
     public AddressBook build() {
         return addressBook;
+    }
+
+    /**
+     * Returns an {@code AddressBook} with all the typical persons and tasks.
+     */
+    public static AddressBook getTypicalAddressBook() {
+        AddressBook ab = new AddressBook();
+        for (Person person : getTypicalPersons()) {
+            ab.addPerson(person);
+        }
+        for (Task task : getTypicalTasks()) {
+            ab.addTask(task);
+        }
+        return ab;
     }
 }
